Do not set CONFIG_IRDA and fix statement on Appletalk
Appletalk actually always needed to be enabled in the kernel. Before 5.17 it just did not cause build errors as netdevice.h defined the required atalk_ptr because CONFIG_IRDA was set but if the kernel was not actually compiled with either CONFIG_IRDA (< 4.18) or CONFIG_ATALK the resulting module did not work.

By not setting CONFIG_IRDA we cause the same build errors on all kernel versions if Appletalk is not enabled instead of building a broken module.